requirements.txt
fastapi
uvicorn
youtube-transcript-api
yt-dlp
python-dotenv
groq
